+++ packages/agent-browser-extension/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @agent-native/agent-browser-extension
 
+## 0.1.215
+
+### Patch Changes
+
+- Updated dependencies [562194a]
+  - @agent-native/core@0.164.25
+
+## 0.1.214
+
+### Patch Changes
+
+- Updated dependencies [14a3f87]
+- Updated dependencies [14a3f87]
+  - @agent-native/core@0.164.24
+
+## 0.1.213
+
+### Patch Changes
+
+- Updated dependencies [b811566]
+  - @agent-native/core@0.164.23
+
 ## 0.1.212
 
 ### Patch Changes