+++ packages/skills/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @agent-native/skills
 
+## 0.2.655
+
+### Patch Changes
+
+- Updated dependencies [562194a]
+  - @agent-native/core@0.164.25
+
+## 0.2.654
+
+### Patch Changes
+
+- Updated dependencies [14a3f87]
+- Updated dependencies [14a3f87]
+  - @agent-native/core@0.164.24
+
+## 0.2.653
+
+### Patch Changes
+
+- Updated dependencies [b811566]
+  - @agent-native/core@0.164.23
+
 ## 0.2.652
 
 ### Patch Changes